+++ packages/core/src/__generate__/accounts/api.ts
@@ -846,6 +846,12 @@ export interface IdentityDTO {
     accountId: string;
     /**
      * 
+     * @type {boolean}
+     * @memberof IdentityDTO
+     */
+    isDeletedAccount: boolean;
+    /**
+     * 
      * @type {string}
      * @memberof IdentityDTO
      */
@@ -886,6 +892,12 @@ export interface IdentityDTO {
      * @memberof IdentityDTO
      */
     otpInitialized?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof IdentityDTO
+     */
+    passwordInitialized?: boolean;
     /**
      * 
      * @type {string}
@@ -940,6 +952,25 @@ export interface InitializeOtpRequest {
      * 
      * @type {string}
      * @memberof InitializeOtpRequest
+     */
+    otpCode?: string;
+}
+/**
+ * 
+ * @export
+ * @interface InitializePasswordRequest
+ */
+export interface InitializePasswordRequest {
+    /**
+     * 
+     * @type {string}
+     * @memberof InitializePasswordRequest
+     */
+    targetAccountId: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof InitializePasswordRequest
      */
     otpCode?: string;
 }
@@ -1174,6 +1205,25 @@ export enum NotificationPayloadDtoLanguageEnum {
     EN = 'EN'
 }
 
+/**
+ * 
+ * @export
+ * @interface NotifyAccessTokenExpirationRequest
+ */
+export interface NotifyAccessTokenExpirationRequest {
+    /**
+     * 
+     * @type {Array<number>}
+     * @memberof NotifyAccessTokenExpirationRequest
+     */
+    remainingExpirationDays?: Array<number>;
+    /**
+     * 
+     * @type {number}
+     * @memberof NotifyAccessTokenExpirationRequest
+     */
+    expirationCheckMinuteRange?: number;
+}
 /**
  * 
  * @export
@@ -2188,6 +2238,44 @@ export const AccountControllerApiAxiosParamCreator = function (configuration?: C
         },
         /**
          * 
+         * @param {InitializePasswordRequest} initializePasswordRequest 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        initializePassword: async (initializePasswordRequest: InitializePasswordRequest, options: any = {}): Promise<RequestArgs> => {
+            // verify required parameter 'initializePasswordRequest' is not null or undefined
+            if (initializePasswordRequest === null || initializePasswordRequest === undefined) {
+                throw new RequiredError('initializePasswordRequest','Required parameter initializePasswordRequest was null or undefined when calling initializePassword.');
+            }
+            const localVarPath = `/api/v2/accounts/initialize-password`;
+            const localVarUrlObj = globalImportUrl.parse(localVarPath, true);
+            let baseOptions;
+            if (configuration) {
+                baseOptions = configuration.baseOptions;
+            }
+            const localVarRequestOptions = { method: 'POST', ...baseOptions, ...options};
+            const localVarHeaderParameter = {} as any;
+            const localVarQueryParameter = {} as any;
+
+
+    
+            localVarHeaderParameter['Content-Type'] = 'application/json';
+
+            localVarUrlObj.query = {...localVarUrlObj.query, ...localVarQueryParameter, ...options.query};
+            // fix override query string Detail: https://stackoverflow.com/a/7517673/1077943
+            delete localVarUrlObj.search;
+            let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
+            localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
+            const needsSerialization = (typeof initializePasswordRequest !== "string") || localVarRequestOptions.headers['Content-Type'] === 'application/json';
+            localVarRequestOptions.data =  needsSerialization ? JSON.stringify(initializePasswordRequest !== undefined ? initializePasswordRequest : {}) : (initializePasswordRequest || "");
+
+            return {
+                url: globalImportUrl.format(localVarUrlObj),
+                options: localVarRequestOptions,
+            };
+        },
+        /**
+         * 
          * @param {InlineObject} inlineObject 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
@@ -2575,6 +2663,19 @@ export const AccountControllerApiFp = function(configuration?: Configuration) {
         },
         /**
          * 
+         * @param {InitializePasswordRequest} initializePasswordRequest 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        async initializePassword(initializePasswordRequest: InitializePasswordRequest, options?: any): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<void>> {
+            const localVarAxiosArgs = await AccountControllerApiAxiosParamCreator(configuration).initializePassword(initializePasswordRequest, options);
+            return (axios: AxiosInstance = globalAxios, basePath: string = BASE_PATH) => {
+                const axiosRequestArgs = {...localVarAxiosArgs.options, url: basePath + localVarAxiosArgs.url};
+                return axios.request(axiosRequestArgs);
+            };
+        },
+        /**
+         * 
          * @param {InlineObject} inlineObject 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
@@ -2747,6 +2848,15 @@ export const AccountControllerApiFactory = function (configuration?: Configurati
         },
         /**
          * 
+         * @param {InitializePasswordRequest} initializePasswordRequest 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        initializePassword(initializePasswordRequest: InitializePasswordRequest, options?: any): AxiosPromise<void> {
+            return AccountControllerApiFp(configuration).initializePassword(initializePasswordRequest, options).then((request) => request(axios, basePath));
+        },
+        /**
+         * 
          * @param {InlineObject} inlineObject 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
@@ -2904,6 +3014,17 @@ export class AccountControllerApi extends BaseAPI {
      */
     public initializeOtp(initializeOtpRequest: InitializeOtpRequest, options?: any) {
         return AccountControllerApiFp(this.configuration).initializeOtp(initializeOtpRequest, options).then((request) => request(this.axios, this.basePath));
+    }
+
+    /**
+     * 
+     * @param {InitializePasswordRequest} initializePasswordRequest 
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     * @memberof AccountControllerApi
+     */
+    public initializePassword(initializePasswordRequest: InitializePasswordRequest, options?: any) {
+        return AccountControllerApiFp(this.configuration).initializePassword(initializePasswordRequest, options).then((request) => request(this.axios, this.basePath));
     }
 
     /**
@@ -3989,6 +4110,44 @@ export const OperationControllerApiAxiosParamCreator = function (configuration?:
         },
         /**
          * 
+         * @param {NotifyAccessTokenExpirationRequest} notifyAccessTokenExpirationRequest 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        notifyAccessTokenExpiration: async (notifyAccessTokenExpirationRequest: NotifyAccessTokenExpirationRequest, options: any = {}): Promise<RequestArgs> => {
+            // verify required parameter 'notifyAccessTokenExpirationRequest' is not null or undefined
+            if (notifyAccessTokenExpirationRequest === null || notifyAccessTokenExpirationRequest === undefined) {
+                throw new RequiredError('notifyAccessTokenExpirationRequest','Required parameter notifyAccessTokenExpirationRequest was null or undefined when calling notifyAccessTokenExpiration.');
+            }
+            const localVarPath = `/api/v2/operation/notify-access-token-expiration`;
+            const localVarUrlObj = globalImportUrl.parse(localVarPath, true);
+            let baseOptions;
+            if (configuration) {
+                baseOptions = configuration.baseOptions;
+            }
+            const localVarRequestOptions = { method: 'POST', ...baseOptions, ...options};
+            const localVarHeaderParameter = {} as any;
+            const localVarQueryParameter = {} as any;
+
+
+    
+            localVarHeaderParameter['Content-Type'] = 'application/json';
+
+            localVarUrlObj.query = {...localVarUrlObj.query, ...localVarQueryParameter, ...options.query};
+            // fix override query string Detail: https://stackoverflow.com/a/7517673/1077943
+            delete localVarUrlObj.search;
+            let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
+            localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
+            const needsSerialization = (typeof notifyAccessTokenExpirationRequest !== "string") || localVarRequestOptions.headers['Content-Type'] === 'application/json';
+            localVarRequestOptions.data =  needsSerialization ? JSON.stringify(notifyAccessTokenExpirationRequest !== undefined ? notifyAccessTokenExpirationRequest : {}) : (notifyAccessTokenExpirationRequest || "");
+
+            return {
+                url: globalImportUrl.format(localVarUrlObj),
+                options: localVarRequestOptions,
+            };
+        },
+        /**
+         * 
          * @param {string} requestId 
          * @param {RejectCoinListingRequestRequest} rejectCoinListingRequestRequest 
          * @param {*} [options] Override http request option.
@@ -4211,6 +4370,19 @@ export const OperationControllerApiFp = function(configuration?: Configuration) 
         },
         /**
          * 
+         * @param {NotifyAccessTokenExpirationRequest} notifyAccessTokenExpirationRequest 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        async notifyAccessTokenExpiration(notifyAccessTokenExpirationRequest: NotifyAccessTokenExpirationRequest, options?: any): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<void>> {
+            const localVarAxiosArgs = await OperationControllerApiAxiosParamCreator(configuration).notifyAccessTokenExpiration(notifyAccessTokenExpirationRequest, options);
+            return (axios: AxiosInstance = globalAxios, basePath: string = BASE_PATH) => {
+                const axiosRequestArgs = {...localVarAxiosArgs.options, url: basePath + localVarAxiosArgs.url};
+                return axios.request(axiosRequestArgs);
+            };
+        },
+        /**
+         * 
          * @param {string} requestId 
          * @param {RejectCoinListingRequestRequest} rejectCoinListingRequestRequest 
          * @param {*} [options] Override http request option.
@@ -4348,6 +4520,15 @@ export const OperationControllerApiFactory = function (configuration?: Configura
          */
         inactivateOrganization(organizationId: string, options?: any): AxiosPromise<void> {
             return OperationControllerApiFp(configuration).inactivateOrganization(organizationId, options).then((request) => request(axios, basePath));
+        },
+        /**
+         * 
+         * @param {NotifyAccessTokenExpirationRequest} notifyAccessTokenExpirationRequest 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        notifyAccessTokenExpiration(notifyAccessTokenExpirationRequest: NotifyAccessTokenExpirationRequest, options?: any): AxiosPromise<void> {
+            return OperationControllerApiFp(configuration).notifyAccessTokenExpiration(notifyAccessTokenExpirationRequest, options).then((request) => request(axios, basePath));
         },
         /**
          * 
@@ -4510,6 +4691,17 @@ export class OperationControllerApi extends BaseAPI {
      */
     public inactivateOrganization(organizationId: string, options?: any) {
         return OperationControllerApiFp(this.configuration).inactivateOrganization(organizationId, options).then((request) => request(this.axios, this.basePath));
+    }
+
+    /**
+     * 
+     * @param {NotifyAccessTokenExpirationRequest} notifyAccessTokenExpirationRequest 
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     * @memberof OperationControllerApi
+     */
+    public notifyAccessTokenExpiration(notifyAccessTokenExpirationRequest: NotifyAccessTokenExpirationRequest, options?: any) {
+        return OperationControllerApiFp(this.configuration).notifyAccessTokenExpiration(notifyAccessTokenExpirationRequest, options).then((request) => request(this.axios, this.basePath));
     }
 
     /**