+++ packages/api-proxy/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.8.2-alpha.5](https://github.com/HAECHI-LABS/henesis-wallet-sdk/compare/v3.8.2-alpha.4...v3.8.2-alpha.5) (2022-03-15)
+
+
+### Features
+
+* add polygon swagger spec ([f4e3883](https://github.com/HAECHI-LABS/henesis-wallet-sdk/commit/f4e3883dfc1e7b7b6f65e7109fc493e8e3b5c451))
+
+
+
+
+
+## [3.8.2-alpha.4](https://github.com/HAECHI-LABS/henesis-wallet-sdk/compare/v3.8.2-alpha.3...v3.8.2-alpha.4) (2022-03-15)
+
+
+### Features
+
+* add polygon to api proxy ([#654](https://github.com/HAECHI-LABS/henesis-wallet-sdk/issues/654)) ([2edd0ab](https://github.com/HAECHI-LABS/henesis-wallet-sdk/commit/2edd0ab2cf9bd07bc88298a606f9e4e1801b0ac3))
+
+
+
+
+
 ## [3.8.2-alpha.3](https://github.com/HAECHI-LABS/henesis-wallet-sdk/compare/v3.8.2-alpha.2...v3.8.2-alpha.3) (2022-02-18)
 
 **Note:** Version bump only for package @haechi-labs/henesis-wallet-api-proxy